refactor: move priority activity message handler to a separate module
